requirements.txt
ollama
gradio
psycopg2
geopandas
matplotlib
pillow